+++ examples/native-example/src/ComponentWithHook.js
@@ -1,6 +1,8 @@
 import React, { useEffect } from 'react';
+import { Platform } from 'react-native';
 
-import { useBlocklyEditor } from '@react-blockly/core';
+import { useBlocklyNativeEditor } from '@react-blockly/core';
+import { WebView } from 'react-native-webview';
 
 import ConfigFiles from './content';
 
@@ -8,17 +10,32 @@ function ComponentWithHook(props) {
   const { workspaceConfiguration, onInject, onChange, onDispose, onError } =
     props;
 
-  const { editorRef, init, state, updateState, updateToolboxConfig } =
-    useBlocklyEditor({
-      workspaceConfiguration,
-      initial: ConfigFiles.INITIAL_XML,
-      onInject,
-      onChange,
-      onDispose,
-      onError,
-    });
+  const {
+    editorRef,
+    init,
+    dispose,
+    state,
+    updateState,
+    updateToolboxConfig,
+    htmlRender,
+    onMessage,
+  } = useBlocklyNativeEditor({
+    workspaceConfiguration,
+    initial: ConfigFiles.INITIAL_XML,
+    onInject,
+    onChange,
+    onDispose,
+    onError,
+    platform: Platform.OS,
+  });
 
   useEffect(() => {
+    return () => {
+      dispose();
+    };
+  }, []);
+
+  const onLoadEnd = () => {
     // if you need to do init later
     testPromise({ timeout: 1000, data: ConfigFiles.INITIAL_JSON }).then(res => {
       init({
@@ -61,15 +78,26 @@ function ComponentWithHook(props) {
 
     // get state
     testPromise({ timeout: 7000 }).then(() => {
-      console.log('state', state());
+      console.log('state', JSON.stringify(state()));
     });
-  }, []);
+  };
 
   function testPromise({ data, timeout = 1000 }) {
     return new Promise(res => setTimeout(() => res(data), timeout));
   }
 
-  return <div className="editor" ref={editorRef}></div>;
+  return (
+    <WebView
+      style={{ flex: 1 }}
+      ref={editorRef}
+      originWhitelist={['*']}
+      source={{
+        html: htmlRender(),
+      }}
+      onMessage={onMessage}
+      onLoadEnd={onLoadEnd}
+    />
+  );
 }
 
 export { ComponentWithHook };